WIP - move couchdb config helpers into separate files
