+++ lib/config/db/couchdb-set-config.js
@@ -1,0 +1,15 @@
+module.exports = setConfig
+
+function setConfig (couch, callback) {
+  couch({
+    url: '/_config/httpd/authentication_handlers',
+    method: 'PUT',
+    body: '{couch_httpd_oauth, oauth_authentication_handler},{couch_httpd_auth, default_authentication_handler},{couch_httpd_auth, cookie_authentication_handler}'
+  }, function (err, res, data) {
+    if (err || (res && res.statusCode !== 200)) {
+      return callback(new Error('Could not set necessary CouchDB config'))
+    }
+
+    callback(null)
+  })
+}
